backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
pydantic==2.5.0
python-dotenv==1.0.0
requests==2.31.0
openai==1.3.7
pandas==2.1.4
scikit-learn==1.3.2
numpy==1.25.2
python-multipart==0.0.6
httpx==0.25.2
websockets==12.0
aiofiles==23.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pytest==7.4.3
pytest-asyncio==0.21.1 